Refactor recovery average
This commit is a bit of a shot in the dark for the issue with null
character data, but it does help in other cases where the recovery dice
roll becomes malformed. It refactors how average recoveries are
calculated to have more failsafes in place, such as falling back to a
d8 if the value is otherwise invalid.